changed routers, solved several bugs with auth tokens and other changes

+++ backend/app/validation/schemas.py
@@ -28,5 +28,5 @@ class UserOut(BaseModel):
     
 class Payload(BaseModel):
     sub: str
-    exp: int
-    iat: int
+    exp: float
+    iat: float
